More work on Squirrel AI scripting. "Wait" now works. Took a while to figure this one out :) It works by suspending the current VM, queueing a 'resume' event, and ignoring all other events until resume is run.
The one situation that needs to be resolved is when an instance script calls into the AI script (and vice versa). Should it interrupt the sleep completely leaving the current scope, or somehow continue. I might have sleep() return a value indicating it was interrupted, and leave it to the script to decide what to do. This behaviour would be similar to Java, that can have an InterruptedException thrown during a sleep().