Update QuestAnglorum.txt
Fixed, changed, and tweaked all quests in Fort Stonehenge, Camelot, Farmer Felps's Farm, Camelot Graveyard, Camelot Chamber and the Militia House. Biggest change was Bandit Connection? which had it's full text restored and a new first Act created to fill the void. Assaulting the Keep's long-standing gender issue with Konin Graves is resolved as well. Also readjusted the level range for the BOUNTY: Steppe-Slimer BB in Camelot so that it matches the other one's level range rather than Ft. Stonehenge's set and changed Basic Duty to require The Militia House quest to be completed before it can be taken as it should have been from the beginning.