Added the first batch of the new quests and went through and checked all mechanics in Bastion. On the way, I added the 'Death' animation to the catapults, fixed the invisible 'target' issue and made some improvements to the catapult firing AI scripts that should make them work without having to choose  the appropriate ability yourself.
So I think this means Bastion is now 'complete'.

Also upgrade CivetWeb which seems to fix that odd crash that happens (I suspect this causes the problems Rivers and Ric both see).

Also spent a lot of time trying to fix zone warping from scripts. The root of the problem is that the instance (and so all quest scripts) are shutdown on warp. If there are events queued for execution, this could cause some crashes when the script VM was accessed again. I think I have fixed this, but it involved lots of little tweaks the event queue and support code, so may have messed something else up. Testing will tell.