Update QuestEarthrise.txt
Tweaked, altered and cleaned up the quests in Old Earthrise even though they will not be used in TAW. Adjusted the Highpoint quests involving the Shards to more accurately fit the lore. Removed the duplicate wording regarding Anglorum in Merlin's quest "Assault: Reclaiming Earthend!" so that the following quest "Travel to Anglorum" no longer sounds like a repeat of words.