Update QuestBastionIsland.txt
Added in the missing starter quest "Welcome to Earth Eternal". This is the Bastion version, not the Mushroom Isle one. Realigned all quest IDs to accommodate the restored quest. Also restored the original quest text to: "Welcome to Earth Eternal", "The Anubian War", and "Anubian Scouts". Tweaked a few other spots to blend everything together.