Implements the server side changes required to allow us to do away with the manual registration key process.
A single key for a particular server can now be generated from the users TAW website account. So the sign up process will now be :-

1. User creates account at http://www.theanubianwar.com
2. User goes to their "Account" page
3. User clicks on "Request Key"
4. User enters their required game account username, password and grove name
5. Server responses with success message
6. User may now start game client and login with their requested details

At now point is any intervention by any admin needed. Neither will we ever run out registration keys.

It is also possible to link your existing account to your TAW site account. This will in the future aid the migration to TAW (and allow the gifts for veterans etc), so it would be preferable to persuade everyone they should link their accounts.

It also gives us an email address associated with a player for future communications.