Implemented the first two books of Liskas set to get a feel of actually implementing them for real. It's thrown up a few tweaks that need making, one of which has been done.
Firstly, the "is usable" previously had to be scripted, this is no longer the case. The creature definitions defHints now has a CDEF_HINT_ITEM_GIVER. When set, the ExtraData contains a list of item IDs that the creature will give (currently ONLY when the player doesn't have them anywhere in their inventory).

Secondly, the "on_use" has to scripted. This is still the case. I made one attempt, but realised more was needed, so disabled that code for the moment and put the handler in the script where I know it works.