Added some mitigation against suspected attack causing resource problems (Sockets/Threads). POST requests are now rejected entirely unless the (rather odd) User-Agent 'ire-3dply(VERSION)' is sent. We now do this from RemoteServerTools.html and the TAW website connections for key importing/password change.
GET requests are now rejected entirely unless the same condition is met, but with the exception of RemoteServerTools.html and the other default files used for localhost installs of the server (news, tips etc).

User-Agent is easily fakeable, but this should stop the drive-by attacks working. It would take someone examining our source or monitoring real client network activity to discover the user agent. That would imply a targeted attack by someone other than a script kiddie. Such a person would not bother with the kind of attacks we are getting as they are pointless for anything other than causing low level disruption.