Cumulative List for the Buff/Timer Bug #469
This is for speed buff timer problems. While client only (the server is tracking timers correctly), this annoying problem was actually caused by the "Global Speed Buff". This was supposed to add a 5% buff to all players and other creatures. It did this on entering any instance, and causes the timers to get messed up, as any existing buff is replaced with this one.

I have removed the global speed buff entirely, and the 5% speed boost is now supplied by altering the global default speed to 52.25, which as exactly the same effect without this hack.