Deal with some bitrot, and start getting things moving again :)
Server support for the experimental improved grove tools.

Some Arm specific fixes. This makes the server runnable on a Raspberry Pi without the weird movement glitches. This was due to a difference between how a signed `integer` is casted to an `unsigned char`. The sign is lost on Arm.

Also replaced usages of `rand()` with C++11 random number generation. This should give a much better uniform distribution for random number generation of different types. It also allows different random number engines and distribution algorithms to be used.